[AWIBOF-4338] move isAccessible variable initialize code

+++ src/mapper/map/map.h
@@ -64,7 +64,7 @@ public:
     Map(int numPages, int pageSize);
     virtual ~Map(void);
 
-    virtual int GetSize(void);
+    virtual uint64_t GetSize(void);
     virtual int GetNumMpages(void);
     virtual char* GetMpage(int pageNr);
     virtual char* GetMpageWithLock(int pageNr);
